new things

+++ src/Data/EncomendaDAO.java
@@ -17,101 +17,86 @@ public class EncomendaDAO {
 
     Connection con;
 
-    public void addEncomenda(Encomenda enc){
+    public void addEncomenda(Encomenda enc) throws Exception{
         con = Connect.connect();
-        if(con!=null){
-            try{
-                PreparedStatement ps = con.prepareStatement("Insert into Encomenda (id,estado,descricao) VALUES (?,?,?)");
-                ps.setInt(1,enc.getId());
-                ps.setString(2,enc.getStatus());
-                ps.setString(3,enc.getDescricao());
-                int a = ps.executeUpdate();
+        if(con!=null) {
 
-            }catch (SQLException e){
-                e.printStackTrace();
-            }
+            PreparedStatement ps = con.prepareStatement("Insert into Encomenda (id,estado,descricao) VALUES (?,?,?)");
+            ps.setInt(1, enc.getId());
+            ps.setString(2, enc.getStatus());
+            ps.setString(3, enc.getDescricao());
+            int a = ps.executeUpdate();
 
 
-        }else Connect.close(con);
+        }
+        else Connect.close(con);
     }
 
 
-    public Encomenda getEncomenda(int id){
+    public Encomenda getEncomenda(int id) throws Exception{
         Connection con = Connect.connect();
         Encomenda enc = new Encomenda();
         ArrayList<LinhaDeEncomenda> l = new ArrayList<>();
 
-        if(con!=null){
-            try{
+        if(con!=null) {
 
-                PreparedStatement ps = con.prepareStatement("Select * from Encomenda where id = ?");
-                ps.setInt(1,id);
-                ResultSet rs = ps.executeQuery();
+            PreparedStatement ps = con.prepareStatement("Select * from Encomenda where id = ?");
+            ps.setInt(1, id);
+            ResultSet rs = ps.executeQuery();
 
-                while(rs.next()){
-                    enc.setId(rs.getInt("id"));
-                    enc.setStatus(rs.getString("estado"));
-                    enc.setDescricao(rs.getString("descricao"));
-                }
-
-                //todas as linhas de encomenda peças
-                ps = con.prepareStatement("select LDE.preco , LDE.quantidade , LDEPeça.idPeca from Encomenda as E  inner join LDEncomenda as LDE on E.id = LDE.idEncomenda inner join LDEPeça on LDE.id = LDEPeça.idLDEncomenda where E.id = ?;");
-                ps.setInt(1,id);
-                rs = ps.executeQuery();
-
-                while(rs.next()){
-                    LinhaDeEncomendaPeca ldp = new LinhaDeEncomendaPeca();
-                    ldp.setQuantidade(rs.getInt("quantidade"));
-                    ldp.setId(rs.getInt("idPeca"));
-                    ldp.setPreco(rs.getFloat("preco"));
-                    l.add(ldp);
-                }
-
-                //todas as linahs de encomenda pacotes
-                ps = con.prepareStatement("select LDE.preco , LDE.quantidade , LDEPacote.idPacote from Encomenda as E  inner join LDEncomenda as LDE on E.id = LDE.idEncomenda inner join LDEPacote on LDE.id = LDEPacote.idLDEncomenda where E.id = ?;");
-                ps.setInt(1,id);
-                rs = ps.executeQuery();
-
-                while(rs.next()){
-                    LinhaDeEncomendaPacote ldp = new LinhaDeEncomendaPacote();
-                    ldp.setQuantidade(rs.getInt("quantidade"));
-                    ldp.setId(rs.getInt("idPacote"));
-                    ldp.setPreco(rs.getFloat("preco"));
-                    l.add(ldp);
-                }
-
-
-                //todas as linhas de encomenda
-                enc.setLinhasDeEncomenda(l);
-
-            }catch (SQLException e){
-                e.printStackTrace();
+            while (rs.next()) {
+                enc.setId(rs.getInt("id"));
+                enc.setStatus(rs.getString("estado"));
+                enc.setDescricao(rs.getString("descricao"));
             }
+
+            //todas as linhas de encomenda peças
+            ps = con.prepareStatement("select LDE.preco , LDE.quantidade , LDEPeça.idPeca from Encomenda as E  inner join LDEncomenda as LDE on E.id = LDE.idEncomenda inner join LDEPeça on LDE.id = LDEPeça.idLDEncomenda where E.id = ?;");
+            ps.setInt(1, id);
+            rs = ps.executeQuery();
+
+            while (rs.next()) {
+                LinhaDeEncomendaPeca ldp = new LinhaDeEncomendaPeca();
+                ldp.setQuantidade(rs.getInt("quantidade"));
+                ldp.setId(rs.getInt("idPeca"));
+                ldp.setPreco(rs.getFloat("preco"));
+                l.add(ldp);
+            }
+
+            //todas as linahs de encomenda pacotes
+            ps = con.prepareStatement("select LDE.preco , LDE.quantidade , LDEPacote.idPacote from Encomenda as E  inner join LDEncomenda as LDE on E.id = LDE.idEncomenda inner join LDEPacote on LDE.id = LDEPacote.idLDEncomenda where E.id = ?;");
+            ps.setInt(1, id);
+            rs = ps.executeQuery();
+
+            while (rs.next()) {
+                LinhaDeEncomendaPacote ldp = new LinhaDeEncomendaPacote();
+                ldp.setQuantidade(rs.getInt("quantidade"));
+                ldp.setId(rs.getInt("idPacote"));
+                ldp.setPreco(rs.getFloat("preco"));
+                l.add(ldp);
+            }
+
+
+            //todas as linhas de encomenda
+            enc.setLinhasDeEncomenda(l);
         }else Connect.close(con);
 
         return enc;
     }
 
     //Map<Descricao, Pair<Id, Categoria>>
-    public Map<String, Pair<Integer, String>> getAllEncomendas(){
+    public Map<String, Pair<Integer, String>> getAllEncomendas() throws Exception{
 
         con = Connect.connect();
         Map<String, Pair<Integer, String>> result = new HashMap<>();
 
-        if(con!=null){
-            try{
-                PreparedStatement ps = con.prepareStatement("Select * from Encomenda");
-                ResultSet rs = ps.executeQuery();
+        if(con!=null) {
+            PreparedStatement ps = con.prepareStatement("Select * from Encomenda");
+            ResultSet rs = ps.executeQuery();
 
-                while(rs.next()){
-                    result.put(rs.getString("estado"),new Pair<>(rs.getInt("id"),rs.getString("descricao")));
-                }
-
-
-            }catch (SQLException e){
-                e.printStackTrace();
+            while (rs.next()) {
+                result.put(rs.getString("estado"), new Pair<>(rs.getInt("id"), rs.getString("descricao")));
             }
-
 
         }else Connect.close(con);
 
@@ -119,39 +104,41 @@ public class EncomendaDAO {
     }
 
 
-    public void removeEncomenda(int id){
+    public boolean removeEncomenda(int id) throws Exception{
 
         con = Connect.connect();
 
-        if(con!=null){
-            try{
+        if(con!=null) {
 
-
-                PreparedStatement ps = con.prepareStatement("Delete from LDEPeça where idLDEncomenda in ( select LD.id from Encomenda as E inner join LDEncomenda  AS LD on E.id = LD.idEncomenda where E.id = ?);");
-                ps.setInt(1,id);
-                ps.executeUpdate();
-
-
-                ps = con.prepareStatement("Delete from LDEPacote where idLDEncomenda in ( select LD.id from Encomenda as E inner join LDEncomenda  AS LD on E.id = LD.idEncomenda where E.id = ?);");
-                ps.setInt(1,id);
-                ps.executeUpdate();
-
-                ps = con.prepareStatement("Delete from LDEncomenda where idEncomenda = ?");
-                ps.setInt(1,id);
-                ps.executeUpdate();
-
-                ps = con.prepareStatement("Delete from Encomenda where id = ?");
-                ps.setInt(1,id);
-                ps.executeUpdate();
-
-
-            }catch (SQLException e){
-                e.printStackTrace();
+            PreparedStatement ps = con.prepareStatement("select * Encomenda where id = ?");
+            ps.setInt(1,id);
+            ResultSet rs = ps.executeQuery();
+            if(rs==null){
+                return true;
             }
+
+            ps = con.prepareStatement("Delete from LDEPeça where idLDEncomenda in ( select LD.id from Encomenda as E inner join LDEncomenda  AS LD on E.id = LD.idEncomenda where E.id = ?);");
+            ps.setInt(1, id);
+            ps.executeUpdate();
+
+
+            ps = con.prepareStatement("Delete from LDEPacote where idLDEncomenda in ( select LD.id from Encomenda as E inner join LDEncomenda  AS LD on E.id = LD.idEncomenda where E.id = ?);");
+            ps.setInt(1, id);
+            ps.executeUpdate();
+
+            ps = con.prepareStatement("Delete from LDEncomenda where idEncomenda = ?");
+            ps.setInt(1, id);
+            ps.executeUpdate();
+
+            ps = con.prepareStatement("Delete from Encomenda where id = ?");
+            ps.setInt(1, id);
+            ps.executeUpdate();
+            return true;
 
 
         }else Connect.close(con);
 
+        return false;
 
     }
